docs: add define-match-expander example
The first example for define-match-expander destructures *syntax lists*, making it hard to tell (at least for me!) what parts are dealing with syntax because of that example or because of the API to define-match-expander.

Adds an example which deals with value structs to hopefully make this a little bit clearer and a bit closer to the "minimal working example" of the macro.

Co-authored-by: Jack Firth <jackhfirth@gmail.com>