racket-index: fix document links for layered installations
When a layer renders new documentation that refers to a previous
layer, the redirection script was not generated as needed to reach
the previous layer.

The documentation currently implements less sharing that would be
possible across layers. For each installation layer, `raco setup`
duplicates the documentation database in "docindex.sqlite", and it
duplicates the search index, and it duplicates the local-redirection
index. That duplication avoids having to join across an an arbitrary
number of SQLite databases or chain the search and local-redirection
indexes.