find-executable-path: search with ".exe" on Windows
On Windows, a program with the given name cannot be found and if it
does not have a file extension, then automatically search again with
".exe" added.

In many cases, a Windows executable has the same name as a Unix
executable, but the executable file name has ".exe". Previously,
`find-executable-path` would require that the ".exe" is included.
So, this improvement avoids further instances of

 (find-system-path (if (eq? 'windows (system-type))
                       "program.exe"
                       "program"))

Although ".exe" is not the only file extension for executables on
Windows, constraining the automatic search to ".exe" should be
effective in most cases, based on experience with the above pattern.
Meanwhile, although it's possible to start a file that has no
extension as an executable (using `system*`, for example), it seems
unlikely that anyone wants to look for a suffixless executable but not
find one that ends in ".exe".

Closes #2129