raco exe: add ".exe" or ".app" instead of replacing an existing suffix
More precisely, this is a change to `embedding-executable-add-suffix`
to make it behave the way that should be expected and the way it's
described in documentation.

This is a backward-incompatible change, but the old behavior is weird,
and it seems unlikely that anyone wanted it or relies on it
(especially given that a suffix is changed only for some platforms and
GUI modes).

Closes #3891