compiler/cm: fix race conditions in parallel-compilation lock
Problem 1:

When a thread T is registered for compiling target Z, a monitor thread
M is created to watch whether T dies, in which case a lock held for Z
should be released. The race condition was in a table mapping Z to M;
if T dies and a replacement T' starts before M discoveres that T has
died, then Z could be mapped to a new M', forgetting M. But when M
discovers that T has died, it would end up terminating M'. The
solution is to map Z+T to M and Z+T' to M'.

This is the problem that the stress test is able to expose.

Problem 2:

A message was sent to the locking manager before a monitor thread M
was created. So, if T dies after sending the message but before
creating M, then the lock is never released.

This one is difficult to provoke, because the opportunity for trouble
is so breif. I couldn't create a test to expose it.

Relevant to #2681