expander: more prompts around module expansion-time evaluation
Make expansion-time evaluation more consistent with run-/visit-time
evaluation in a module by adding prompts around `for-syntax`
definitions and expressions. It's difficult to imagine that the new
prompts are all that useful, but maybe consistency is its own reward.

As an example, this module now expands without error:

 #lang racket/base
 (require (for-syntax racket/base))
 (begin-for-syntax
   (define k (call/cc (lambda (k) k))) ; capture delimited
   (set! k k) ; make `k` mutable
   (k 5) ; apply captured

   (define k2 #f)
   ;; capture in an expression (instead of a definition):
   (let ([k (call/cc (lambda (k) k))])
     (set! k2 k))
   (k2 6))

The example works because the continuation is delimited around each
for-syntax definition and expression. Formerly, the continuations were
delimited only after expansion when the module is visited.

Related to #2617