regexp: fix interaction of lookbehind and prefix bytes
A lookbehind pattern can include a group, and when prefix bytes are
supplied for matching, then the lookbehind group can span bytes that
are in the prefix. That has various subtle implications, such as the
fact that `regexp-match-positions` can return negative indices after
the first pair, and `regexp-match` needs to be prepared to extract
bytes and strings from the prefix. A group can even span the prefix
and input proper:

 > (regexp-match #rx"(?<=(..))" "a" 0 #f #f #"x")
 '("" "xa")