re-enable more primitive implementation of `regexp-replace*`
Both BC and CS had an internal implementation of `regexp-replace*`,
although each with correctness and/or performance issues. This commit
repairs those problems and uses the more primitive `regexp-replace*`
implementations when the range arguments specify the full input.
Probably the only benefit is that the inpt argument can be returned
unmodified if it contains no matches.

Closes #3404