add syntax-object argument to `module-path-index-resolve`
This extra argument is passed on the module name resolver, if the
resolve is used. Within the expander, it's used to communicate a
source location for a `require` to the resolver, but it makes sense to
expose the argument more generally in `module-path-index-resolve`.

(cherry picked from commit 32cb175100c19cddbcc810a93e5e94bb365757af)