expander: fix require shadowing and binding spaces
The intent with binding spaces is that a local binding in the default
space "shadows" bindings in non-default spaces by making references to
them ambiguous.

Unfortunately, shadowing at the level of module bindings (`define`
shadowing a `require` or a `require` shadowing the initial import)
doesn't use layers of scopes. (Why not? Possibly for historical
reasons, and probably it was a mistake. I'm not sure.) So, the effect
of default-space bindings on non-default-space bindings at the module
level has to be replicated specially.

This commit also fixes a few smaller bugs with binding spaces that
were revealed through more tests.