cs & thread: fix problem in thread scheduler
When a `sync` is polled while other threads run, and then the other
threads go away after the `sync` returns, it was possible for the
scheduler to be left in a state where it thought all threads had
nothing to do after the most recent poll attempt, in which case it
would go to sleep. The problem was easiest to trigger in a non-main
place, just because less tends to be going on there.