rerequire: fix reloading for diamond dependencies
With a depedency graph like

  a -> b
  a -> c
  b -> d
  c -> d

where changes were made to `d`, say `check-latest` checked the
dependencies in this order:

  a b d c d

while checking `b`, it would reload `d` and mark it as loaded, then
reload `b` and move on to `c`.  At `c`, it would check `d` again, but
because it had already marked `d` as loaded, it would skip `d` so `c`
wouldn't get reloaded either.

This change fixes that problem by reusing the reloaded cache to
determine whether or not a dependency of some module was already
reloaded.