git-checkout: fix commit ID search when there's no "master" branch
When looking for a commit by ID, a search would try the "master"
branch, but that attempt failed if the repository has no such branch.
Fix that failure (but continuing the search), but also add a
`git-checkout` argument that's a branch or tag name to try first
instead of "master".

This repair fixes a problem with package URLs that end in a fragment
other than "#master" in the (ephemeral) case that the package server
hasn't caught up to the latest commit in the branch/tag and the
branch/tag "master" does not exist.