raco setup: improved support for cross-compile mode
Originally, cross-compilation support in `raco setup` (and `raco
make`) was intended for building distributions, so it could rely on
things being not too out-of-sync. In particular, an up-to-date
compiled form for the host machine would mean that the cross-compiled
form for the target machine must already be in place.

The `raco cross` tool exposes `raco setup` and `raco make` in
cross-compilation mode to more settings where distribution-build
assumptions may not hold. For example, someone might delete "compiled"
directories that hold cross-compiled forms without deleting the
host-compiled form. To handle more of those settings, add triggers for
recompilation to `raco setup` and `raco make` based on extra checks
for the cross-compiled form.

Related to racket/raco-cross#5