cs: fix ordered finalization in GCed custodian
Commit 64f0c8a7ed exposed a problem with custodians and caused a leak
in document renrdering for a distribution --- the jump in the blue
region in https://build-plot-cs.racket-lang.org/2021-05-23.html
relative to https://build-plot-cs.racket-lang.org/2021-05-22.html

The problem was related to registering in a custodian that itself
becomes unreachable and is merged into its parent. In that case, the
merging could create a new late will executor for ordering and
register a value, which is bad if a finalization was meant to be
registered after a custodian registration. The repair is to do less
work and carry over a wrapped callback from the old custodian to the
new.