Add `file-or-directory-stat`
Add `file-or-directory-stat` to `racket/src/io/file/main.rkt`. The new
code is copied and modified from `file-or-directory-identity`. When
`file-or-directory-stat` works, both functions should be consolidated.
This could even mean that `file-or-directory-identity` is pure Racket
code that uses `file-or-directory-stat` to get the necessary data.

Refactoring should start if we have (better) test coverage for both
`file-or-directory-identity` and `file-or-directory-stat`.