syntax-local-eval: accept S-expressions
Revised the change in 8cdab1bf5c to not include a check that the
expression to evaluate is a syntax object. Whether or not it's a good
idea, existing code (at least in the "dssl2" package) relies on being
able to provide a value that gets implicitly coerced to a syntax
object.