Include timestamp values in `stat` result
Note that the C struct field `st_*tim.tv_nsec` contains the nanoseconds
as a fraction of the seconds, i. e. the seconds are _not_ included here.
This makes sense on the C level to avoid overflowing a 64-bit integer
with the digits for seconds _and_ nanoseconds.

On the other hand, since Racket doesn't have a limitation on the digits
in an integer, it's more convenient to combine the seconds and
nanoseconds parts of a timestamp in a single integer value. For example,
seconds = 1631990816 and nanoseconds = 677671045 becomes the combined
nanoseconds value 1631990816677671045.

I also changed the C type from `intptr_t` to `uintptr_t` since these
values are always >= 0.