Change `permission-bits` to `mode`
The term "permission bits" just doesn't apply because the stat "mode"
contains more information. Using the term "permission bits" is rather
likely to lead to confusion with the return value of
`(file-or-directory-permissions path 'bits)`, which has some of the bits
of the "mode" masked out.

Attempt to explain the "mode" in the Reference Guide by referring to
the C `stat` and `lstat` functions.