Handle change/creation time dualism
On Posix, the `st_ctime` field of `stat` structures is the last status
change time. However, on Windows, the `ctime` is the creation time.

Therefore, rename the `change_time_seconds` and
`change_time_nanoseconds` of the `rktio_stat_t` struct field to
`ctime_seconds` and `ctime_nanoseconds`, respectively. Also add a flag
to the struct that tells whether `ctime` refers to the change time or
the creation time.

In `racket/src/io/file/main.rkt`, check the flag and populate the
returned hash with all four keys `'change-time-seconds`,
`'creation-time-seconds`, `'change-time-nanoseconds` and
`'creation-time-nanoseconds`, but set the unavailable values to 0.

Another approach could be to set the missing values to `#f`, but as I
understand it, Windows will use the value 0 if some data isn't available
for a file system, for example the creation time on FAT file systems.
Since we don't want to duplicate any checks that `wstat64` does, it's
handy to just use the 0 values from Windows. And in that case, it
doesn't make sense to have two different values (`#f` vs. 0) for missing
values, so stick to the common value 0.