makefile: avoid `codesign` on Mac OS in favor of Racket implementation
It seems that using `codesign` to remove a signature might sometimes
leave the file in a state that `strip` does not like. If true, that
behavior might be related to build-tool changes around Monterey. In
any case, Racket's version needs to work on executables derived from
the `racket` excutable and launcher, so let's use it always.