Add Windows code for `file-or-directory-stat`
Since I can't build Racket under Windows, I can't test and fully
integrate the Windows code here.

My Windows code is a mixture of a small standalone program I compiled
and executed in a Windows VM on the one hand and some snippets of code I
found for other functions in `rktio_fs.c` for Windows on the other hand.

Matthew, in particular please carefully check

- Path conversion to wchar type
- Error handling

But watch out for other problems as well that I might have overlooked.
;-)

The sections for Posix and Windows that copy values from `stat_buf` to
`rktio_stat_buf` are similar, but not the same:

- Some comments refer only to Posix, some only to Windows.
- The struct fields for timestamps in `stat_buf` are different for
  Posix and Windows.

Though it would be possible to put common code lines in a section
outside the `ifdef`s, it could well be that this refactoring makes the
code _more_ difficult to follow. Also keep in mind that we just copy
data from stat structs that are unlikely to change, due to backward
compatibility considerations. So we might not benefit from being able to
change common code in one place. Also it might be that the "common" code
needs to be changed in different ways for Posix and Windows.

Long story short: Probably it's better to keep the Posix and Windows
implementations separate, but I leave it to you.