expander: add portal syntax with `#%require` support
Portal syntax provides a way to represent and extract binding from a
module at any phase, but without requiring bindings and runtime
support at the phase where it's bound, and without requiring that a
module is instantiated to provide information at that phase.

Portal syntax helps implement importing from modules with hierarchical
names, which is a feature of a current Rhombus prototype. The
essential problem is binding a name at each phase N where a module
import provides bindings at phase N. N might not be 0, either because
the import is shifted to phase N with `for-meta`, or because the
providing module exports bindings a phase N. Either way, a
hierarchical name needs to be bound at phase N. Creating a
`define-syntax` at phase N may be invenient (e.g., at a much higher
phase level than the `require`, or at a negative phase that would
require a submodule to reach) or impossible (at the label phase).
Portal syntax provides a simple and direct way to create that binding.

In other words, portal syntax enables something like

 (define-for-meta <phase> <id> (quote-syntax <stx>))

Adding `define-for-meta` as a new module- and top-level form, however,
would create trouble for many existing tools and macros that work on
fully expanded modules or that partially expand module bodies (such as
a `#%module-begin` implementation). To reduce compatibility problems,
the new definition form is tucked away inside `#%require`:

 (#%require (for-meta <phase> (portal <id> <stx>)))

This avoids problems because few tools attempt to inspect a
`#%require` expansion. The handful of tools that do inspect
`#%require` expansions, such as Check Syntax or DrComplete, will need
to be updated. (This was also true of the recent `for-space`
addition.)

The new pieces:

 * `make-portal-syntax`, `portal-syntax?`, and `portal-syntax-content`:
   a new primitive value. It's is somewhat like a rename transformer
   in that it's normally used at compile time and bound to an
   identifier, potentially via `define-for-syntax`.

 * `(portal <id> <stx>)` subform in `#%provide`: binds <id> to
   portal syntax whose content is <stx>.

   The binding of `<id>` is like one that could be written as
   `(define-syntax <id> (make-portal-syntax ....))`, but the `portal`
   form is able to bind `<id>` in any phase, including the label phase
   or a template phase.

 * `identifier-binding-portal-syntax`: like `syntax-local-value`
   composed with `portal-syntax-content`, but can be used at any time,
   and does not require instantiating the module that binds an
   identifier to glue syntax. The resulting syntax is phase shifted in
   a way that is consistent with the module being instantiated at the
   relevant phase (if it were to be instantiated), even at the label
   phase. The "no instantiation" part here is particularly related to
   using label-phase bindings for tasks like documentation linking.

 * `identifier-distinct-binding`: not really tied to portal syntax,
   but a handy shortcut to checking whether a scope is relevant to
   finding a binding, instead of comparing the results of tro calls to
   `identifier-binding`.

Constraining portal bindings to syntax objects, as opposed to allowing
a generative structure type, ensures that getting portal-syntax
content from an uninstantiated module makes sense. Binding information
on a "tag" in a portal-syntax target might be used to distinguish
different kinds of portal syntax.

A language with hierarchical naming can use these pieces to represent
some (or all) hierarchical names with portal-syntax bindings. Then, it
gets the expander's fancy support for representing bindings
efficiently as well as a way to use template phases and the label
phase.