Generalize for/list and for*/list to work with multiple values
This PR allows each iteration of for/list to accumulate more or less than one element.

The PR is in the same spirit as #2483, but it simply generalizes the
existing for/list rather than creating an entirely new macro.

The PR is backward compatible for #lang racket, and there is no performance
degradation. Existing uses of for/list, which produces only one value in
the body, will generate the same code (compared to before the PR) after
CP0 performs inlining.

The PR requires Typed Racket to be updated, however.