cs: avoid unending "attempt ... in atomic mode" errors
When there's an attempt to block/descheulde a thread in atomic mode or
that is already descheduled, then we're in bad state and there are no
guarantees anymore. Still, an unending stream of "attempt ... in
atomic mode" or "tried to deschedule a descheduled thread" errors is
not so nice. Disable atomic mode and quiet descheduling errors when
the corresponding bad state occurs, in the hope that some information
gets reported to the user/program and they can bail out more
gracefully.