[DRAFT] make-temporary-{file,directory}{,*}: add #:permissions
This commit depends on https://github.com/racket/racket/pull/3870
and a hypothetical `#:permissions` argument to `copy-file`,
where `(copy-file #:permissions #f . args)` would be the default
and preserve the current behavior, i.e. copying the permissions
of the source file.

For more thoughts on changes to `copy-file`, see:
<https://github.com/racket/racket/pull/3870#issuecomment-957971651>
and
<https://github.com/racket/racket/pull/3870#issuecomment-1006990869>.

Before merging, it would also need to add tests.